Prevent do_start_receive when in closing state
When we send ourself an empty packet because all
sockets stopped receiving, we're in a special state
when we're receiving but have the receive_calls variable
set to zero. During this time user can invoke
async_read which would start a parallel read operation,
to avoid that I added another atomic variable to
indicate we're in this limbo state.